47: extract the "email-address-related" test cases (most of which have been introduced in this feature branch) out into their own module
